Update ESLint and Rubocop configs (#2)
Both of these were slightly incompatible with the current versions of
these tools.  I used the warning outputs from Code Climate to identify
these, and many of the warnings suggested what to replace the deprecated
settings with.

I also enabled a few Rubocop style checkers for whitespace checks that I
think will not be controversial.